+++ data/backend/backend/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for ecobalyse project.
+Django settings for backend project.
 
 Generated by 'django-admin startproject' using Django 5.0.3.
 
@@ -36,7 +36,7 @@ ALLOWED_HOSTS = [
 # Application definition
 
 INSTALLED_APPS = [
-    "backend.apps.BackendConfig",
+    "textile.apps.TextileConfig",
     "django.contrib.admin",
     "django.contrib.auth",
     "django.contrib.contenttypes",
@@ -55,7 +55,7 @@ MIDDLEWARE = [
     "django.middleware.clickjacking.XFrameOptionsMiddleware",
 ]
 
-ROOT_URLCONF = "ecobalyse.urls"
+ROOT_URLCONF = "backend.urls"
 
 TEMPLATES = [
     {
@@ -73,7 +73,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = "ecobalyse.wsgi.application"
+WSGI_APPLICATION = "backend.wsgi.application"
 
 
 # Database
