+++ data/backend/reinit.sh
@@ -4,4 +4,4 @@ rm -f backend/migrations/*
 python manage.py makemigrations backend
 python manage.py migrate
 echo "Creating initial data"
-python manage.py shell -c "from backend.init import init; init()"
+python manage.py shell -c "from textile.init import init; init()"
